refactor(guest-agent): split checkpoint pipelines
